# [not for landing] proof of concept that global map is just fine