not really for review - but happy to make it an actual PR

There's a toy benchmark included. The number of entries in the map grows to 10k, but there's no visible impact on performance. I'm not sure how to best benchmark this code - suggestions are welcomed.

Submitting here to hit CI