@bddppq 2018-05-31
There will be perf tests running in https://ci.pytorch.org/jenkins/job/pytorch-builds/job/pytorch-linux-xenial-cuda8-cudnn6-py3-trigger/9672/