@ezyang 2018-05-31
Benchmarks I would do:
- The timeit in https://github.com/pytorch/pytorch/issues/5388
- The dynet benchmark in https://github.com/pytorch/pytorch/issues/2518