@ezyang 2018-05-31
Hi @dzhulgakov, here is the timeit for indexing. Test ran on devfair024, I didn't do any CPU/NUMA pinning. CPU governor is on performance. Intel(R) Xeon(R) CPU E5-2650 v4 @ 2.20GHz.

Baseline fce6b24468cfb29d1a6566772215195b92a50e8e

```
In [1]: import torch
x
In [2]: x = torch.randn(1000)

In [4]: %timeit x[0]
1.11 µs ± 1.34 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: %timeit x[0]
1.11 µs ± 1.54 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [6]: %timeit x[0]
^[[A^[[A1.12 µs ± 7.03 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [7]: %timeit x[0] = 1
3.74 µs ± 12.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [8]: %timeit x[0] = 1
3.75 µs ± 5.55 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Your PR 04ace5f94f625fe9ac6d8e750406d66997db8dc7

```
In [1]: import torch
x
In [2]: x = torch.randn(1000)

In [3]: %timeit x[0]
1.44 µs ± 1.99 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit x[0]
1.37 µs ± 1.96 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: %timeit x[0]
1.38 µs ± 6.61 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [6]: %timeit x[0]
1.38 µs ± 20.6 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [7]: %timeit x[0] = 1
4.3 µs ± 1.24 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [8]: %timeit x[0] = 1
4.28 µs ± 9.39 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [9]: %timeit x[0] = 1
4.32 µs ± 5.88 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Taking the lowest number of all runs:

```
      baseline  change  % diff
get   1.11      1.37    23%
set   3.74      4.28    14%
```

This seems quite appreciable.