@dzhulgakov 2018-05-31
Thanks for the pointers to benchmarks.

From run on my devgpu:

before:
get 1.02us
set 4.17us
dynet 4.131s (not stat-sig)

after:
get 1.13us
set 4.3us
dynet 4.132s (not stat-sig)

dynet benchmark: https://gist.github.com/dzhulgakov/bd53c5b1cbac795321c30a51f92b2d72 (tweaked a bit)
ran with `time OMP_NUM_THREADS=1 MKL_NUM_THREADS=1 python mini_sequence_labeler.py` (with mkl time is higher and variability is huge)
governor is on performance
variability is still quite high for dynet one

So I'd take the microbench with a grain of salt on how much it impacts actual perf