@soumith 2018-05-31
also, once we get into recursive nets, it is a thing to have graphs with 50k nodes and stuff. I remember @jekbradbury reporting some of their internal workloads in that range. Do you think we can up the benchmark to see 100k live Tensors (assuming 2 Tensors live per node), instead of 10k.