@goldsborough 2018-05-31
Note that `unordered_map` is as shit as a hashtable gets. I wouldn't discredit this approach until you actually try a fast hashtable. [`google::dense_hash_map`](https://github.com/sparsehash/sparsehash) can be up to 100 times faster than `unordered_map` for lookups. See P57476999 (fb internal) for exact benchmarks