@soumith 2018-05-31
@goldsborough can you post the benchmarks in an external accessible place. gist.github.com works fine.