@ezyang 2018-06-01
Before I get into more numbers, I want to preface this with more discussion about implicit PyTorch needs, which explain the intuitive aversion of many developers to using a hash table here.

* Performance of wrapping tensors in PyTorch is critical for CPU NLP models in PyTorch, c.f. https://github.com/gchanan/pytorch/wiki/Mini-Sequence-Labeler-PyTorch-vs-Dynet (h/t @dzhulgakov for this text)
* We want simple, predictable behavior in all cases. If something is simple and predictable, there is no chance that it will blow up in our face in some weird, esoteric condition. A single field in the C++ object is simple and predictable. A hash table is not simple and it is only mostly predictable: there are many footguns like poor choice of hash function/distribution of hashes, hash table load factor, cache behavior, probing/collision strategy, which all need to be handled. This is not to say a well written hash table cannot be made to work well in all cases, but if something goes wrong in some esoteric environment, we are far more likely to be able to figure it out if you have a simple implementation strategy instead of a complex one.Case in point, we would not have to do the analysis below if we were using a simple strategy; with a hash table, we have to think and benchmark. There is definite merit to avoiding thinking/benchmarking if it is not necessary.

----

@colesbury requested information on the absolute overheads of the hash table lookup. `ska::flat_hash_map` is described in more detail at https://probablydance.com/2017/02/26/i-wrote-the-fastest-hashtable/ (though, unfortunately, he does not say what hardware he did his tests on.) @soumith mentioned that 100k live tensors is a workload we see in practice. Cross-referencing the blog post, and simply "using my eyeball" to compute what the extra numbers are, we see that we are looking at **successful lookup in the range of 8-10ns** (depending on if we are using prime numbers of power of two; I am not sure if we are allowed to use the power of two resizing policy with integer pointer hashes--an example of complexity) and you are just barely in the range where the hash table fits in cache; **successful lookup with cache miss will be more like 25-30ns** (this is about the same cost of a mutex lock/unlock (citing https://gist.github.com/hellerbarde/2843375#gistcomment-1896153), and Python function call overhead is something like 150ns (citing https://ilovesymposia.com/2015/12/10/the-cost-of-a-python-function-call/)). **Unsuccessful lookup with cache miss will be 30-40ns.** And **insertion is 50-70ns** and **deletion 20-30ns**.

Let's discuss the average lifecycle of entries in this table. It will look something like this:

1. A tensor gets created in C++
2. It moves to Python, at which point an unsuccessful lookup and subsequent insert occurs. (The cost here is rolled up with the cost of allocating the initial PyObject to represent the tensor.)
3. *Possibly, some number of times*, another movement to Python happens, at which point a successful lookup occurs.
4. At some point, the Python object goes dead, and the destructor deletes the object from the hash table. We possibly go back to (2)

Given this lifecycle, `ska::flat_hash_map` is described as being optimized for fast lookup, may not be optimal for PyTorch, where most objects crossing the C-Python FFI boundary are new objects. To assess this, I added counters for the "Wrapping var, existing/new one" debug prints from @dzhulgakov's implementation and ran a single epoch of training with `word_language_model` on CUDA from https://github.com/pytorch/examples. I found that there were 138839 hits to existing variables versus 545558 fresh variables.  To be honest, that is more existing variable returns than I expected, but you can see that we are going to be doing a lot of inserting.

(more to come)